Move AuthMiddleware into socket
And use socket middlewares from jobapi routes
